Cleanup: move "useFindEntityAndVariable" to "study" hooks; delete "studyMetadata" hooks
