+++ src/lib/core/components/visualizations/implementations/MosaicVisualization.tsx
@@ -16,7 +16,7 @@ import {
   MosaicRequestParams,
 } from '../../../api/data-api';
 import { usePromise } from '../../../hooks/promise';
-import { useFindEntityAndVariable } from '../../../hooks/studyMetadata';
+import { useFindEntityAndVariable } from '../../../hooks/study';
 import { useDataClient, useStudyMetadata } from '../../../hooks/workspace';
 import { Filter } from '../../../types/filter';
 import { PromiseType } from '../../../types/utility';
