+++ src/lib/core/components/visualizations/implementations/ScatterplotVisualization.tsx
@@ -19,7 +19,7 @@ import {
 } from '../../../api/data-api';
 
 import { usePromise } from '../../../hooks/promise';
-import { useFindEntityAndVariable } from '../../../hooks/studyMetadata';
+import { useFindEntityAndVariable } from '../../../hooks/study';
 import { useDataClient, useStudyMetadata } from '../../../hooks/workspace';
 import { Filter } from '../../../types/filter';
 import { StudyEntity } from '../../../types/study';
