+++ src/lib/core/hooks/study.ts
@@ -6,9 +6,16 @@ import {
   getScopes,
   getNodeId,
 } from '@veupathdb/wdk-client/lib/Utils/CategoryUtils';
-import { StudyMetadata, StudyRecordClass, StudyRecord } from '../types/study';
+import {
+  StudyEntity,
+  StudyMetadata,
+  StudyRecordClass,
+  StudyRecord,
+} from '../types/study';
 import { usePromise } from './promise';
 import { SubsettingClient } from '../api/subsetting-api';
+import { Variable } from '../types/variable';
+import { findEntityAndVariable } from '../utils/study-metadata';
 
 const STUDY_RECORD_CLASS_NAME = 'dataset';
 
@@ -88,5 +95,12 @@ export function useStudyMetadata(datasetId: string, store: SubsettingClient) {
         );
       return store.getStudyMetadata(study.id);
     }, [datasetId, store])
+  );
+}
+
+export function useFindEntityAndVariable(entities: StudyEntity[]) {
+  return useCallback(
+    (variable?: Variable) => findEntityAndVariable(entities, variable),
+    [entities]
   );
 }
